input processing and file reading added

+++ list/Item.java
@@ -2,10 +2,10 @@ package list;
 
 public class Item {
     private String item_name;
-    private int inventory_count;
-    private int value;
+    private double inventory_count;
+    private double value;
 
-    public Item(String item_name, int value, int inventory_count) {
+    public Item(String item_name, double value, double inventory_count) {
         this.item_name = item_name;
         this.value = value;
         this.inventory_count = inventory_count;
@@ -16,19 +16,19 @@ public class Item {
         return item_name;
     }
 
-    protected int get_Value() {
+    protected double get_Value() {
         return value;
     }
 
-    protected int get_Inventory() {
+    protected double get_Inventory() {
         return inventory_count;
     }
 
-    protected void set_Inventory(int count) {
+    protected void set_Inventory(double count) {
         inventory_count = count;
     }
 
-    protected void set_price(int price) {
+    protected void set_price(double price) {
         value = price;
     }
