+++ list/Manager.java
@@ -8,7 +8,7 @@ public class Manager {
         main_list = new ItemList();
     }
 
-    public void create(String item_name, int price, int count) {
+    public void create(String item_name, double price, double count) {
         main_list.sorted_add(new Item(item_name, price, count));
     }
 
@@ -16,7 +16,7 @@ public class Manager {
         main_list.remove(to_delete);
     }
 
-    public void edit_inventory(String item_name, int new_count) {
+    public void edit_inventory(String item_name, double new_count) {
         Item to_edit = main_list.search(item_name);
 
         if (to_edit != null) {
@@ -32,9 +32,8 @@ public class Manager {
         }
     }
 
-    public void edit_price(String item_name, int new_price) {
+    public void edit_price(String item_name, double new_price) {
         Item to_edit = main_list.search(item_name);
-
         if (to_edit != null) {
             to_edit.set_price(new_price);
         }
@@ -43,5 +42,4 @@ public class Manager {
     public void print_list() {
         main_list.print_list();
     }
-
 }
